+++ examples/hello_world.rs
@@ -16,12 +16,15 @@ fn main() -> wry::Result<()> {
   let window = WindowBuilder::new()
     .with_title("Hello World")
     .build(&event_loop)?;
-  let webview = WebViewBuilder::new(window)?
-    .with_url("https://html5test.com")?
-    .build()?;
+  let webview = WebViewBuilder::new(window)?.with_url("https://html5test.com")?;
 
   #[cfg(debug_assertions)]
-  webview.devtool();
+  let webview = webview.with_devtools(true);
+
+  let webview = webview.build()?;
+
+  #[cfg(debug_assertions)]
+  webview.open_devtools();
 
   event_loop.run(move |event, _, control_flow| {
     *control_flow = ControlFlow::Wait;