+++ src/lib.rs
@@ -19,7 +19,7 @@
 //!     },
 //!     webview::WebViewBuilder,
 //!   };
-//!  
+//!
 //!   let event_loop = EventLoop::new();
 //!   let window = WindowBuilder::new()
 //!     .with_title("Hello World")
@@ -52,15 +52,15 @@
 //! interacting with the window. Enabled by default.
 //! - `protocol`: Enables [`with_custom_protocol`] to define custom URL scheme for handling tasks like
 //! loading assets. Enabled by default.
-//! - `tray`: Enables system tray and more menu item variants on **Linux**. You can still create
-//! those types if you disable it. They just don't create the actual objects. We set this flag
-//! because some implementations require more installed packages. Disable this if you don't want
-//! to install `libappindicator` package. Enabled by default.
-//! - `ayatana`: Enable this if you wish to use more update `libayatana-appindicator` since
-//! `libappindicator` is no longer maintained.
-//! - `devtool`: Enable devtool on **macOS** requires calling private functions. While it's still
-//! enabled in **debug** build on mac, it will require this feature flag to actually enable it in
-//! **release** build. Avoid this in release build if your app needs to publish to App Store.
+//! - `tray`: Enables system tray and more menu item variants on **Linux**. This flag is enabled by default.
+//!  You can still create those types if you disable it. They just don't create the actual objects. We set this flag because some implementations require more installed packages.
+//! - `ayatana-tray`: Enable this if you wish to use more update `libayatana-appindicator` since `libappindicator` is no longer maintained.
+//!  This flag is enabled by default. Disable this if you don't want to install the `libayatana-appindicator` package.
+//! - `gtk-tray`: Enable this if you wish ot use `libappindicator` for tray on **Linux**. The package is supported on more Linux distributions, but it is not maintained anymore.
+//!  Note that `ayatana-tray` and `gtk-tray` cannot be enabled at the same time, so `default-features` must be set to `false`.
+//! - `devtools`: Enables devtools on release builds. Devtools are always enabled in debug builds.
+//! On **macOS**, enabling devtools, reuires calling private apis so you should enabling this flag in release
+//! build if your app needs to publish to App Store.
 //! - `transparent`: Transparent background on **macOS** requires calling private functions.
 //! Avoid this in release build if your app needs to publish to App Store.
 //! - `fullscreen`: Fullscreen video and other media on **macOS** requires calling private functions.