+++ src/webview/android/mod.rs
@@ -1,4 +1,4 @@
-use std::{ffi::c_void, ptr::null_mut, rc::Rc, sync::RwLock};
+use std::{collections::HashSet, ffi::c_void, ptr::null_mut, rc::Rc, sync::RwLock};
 
 use crate::{application::window::Window, Result};
 
@@ -36,7 +36,16 @@ impl InnerWebView {
 
   pub fn focus(&self) {}
 
-  pub fn devtool(&self) {}
+  #[cfg(any(debug_assertions, feature = "devtools"))]
+  pub fn open_devtools(&self) {}
+
+  #[cfg(any(debug_assertions, feature = "devtools"))]
+  pub fn close_devtools(&self) {}
+
+  #[cfg(any(debug_assertions, feature = "devtools"))]
+  pub fn is_devtools_open(&self) -> bool {
+    false
+  }
 
   pub fn run(self, env: JNIEnv, _jclass: JClass, jobject: JObject) -> Result<jobject> {
     let string_class = env.find_class("java/lang/String")?;
@@ -48,21 +57,45 @@ impl InnerWebView {
     // )?;
     let WebViewAttributes {
       url,
+      custom_protocols,
       initialization_scripts,
       ipc_handler,
+      devtools,
       ..
     } = self.attributes;
 
-    // todo
-    // ipc too?
     if let Some(i) = ipc_handler {
       let i = UnsafeIpc(Box::into_raw(Box::new(i)) as *mut _);
       let mut ipc = IPC.write().unwrap();
       *ipc = i;
     }
 
+    if devtools {
+      #[cfg(any(debug_assertions, feature = "devtools"))]
+      {
+        let class = env.find_class("android/webkit/WebView")?;
+        env.call_static_method(
+          class,
+          "setWebContentsDebuggingEnabled",
+          "(Z)V",
+          &[devtools.into()],
+        )?;
+      }
+    }
+
     if let Some(u) = url {
-      let url = env.new_string(u)?;
+      let mut url_string = String::from(u.as_str());
+      let schemes = custom_protocols
+        .into_iter()
+        .map(|(s, _)| s)
+        .collect::<HashSet<_>>();
+      let name = u.scheme();
+      if schemes.contains(name) {
+        url_string = u
+          .as_str()
+          .replace(&format!("{}://", name), "https://tauri.wry/")
+      }
+      let url = env.new_string(url_string)?;
       env.call_method(jobject, "loadUrl", "(Ljava/lang/String;)V", &[url.into()])?;
     }