+++ src/webview/webview2/mod.rs
@@ -94,26 +94,28 @@ impl InnerWebView {
 
     CreateCoreWebView2EnvironmentCompletedHandler::wait_for_async_operation(
       Box::new(move |environmentcreatedhandler| unsafe {
-        if let Some(data_directory) = data_directory {
-          // If we have a custom data_directory, we need to use a call to `CreateCoreWebView2EnvironmentWithOptions`
-          // instead of the much simpler `CreateCoreWebView2Environment`.
-          let options: ICoreWebView2EnvironmentOptions =
-            CoreWebView2EnvironmentOptions::default().into();
-          let data_directory = pwstr_from_str(&data_directory);
-          let result = CreateCoreWebView2EnvironmentWithOptions(
-            PWSTR::default(),
-            data_directory,
-            options,
-            environmentcreatedhandler,
-          )
-          .map_err(webview2_com::Error::WindowsError);
-          let _ = take_pwstr(data_directory);
+        let options: ICoreWebView2EnvironmentOptions =
+          CoreWebView2EnvironmentOptions::default().into();
 
-          return result;
-        }
+        // remove "mini menu" - See https://github.com/tauri-apps/wry/issues/535
+        let additional_arguments = pwstr_from_str("--disable-features=msWebOOUI,msPdfOOUI");
+        let _ = options.SetAdditionalBrowserArguments(additional_arguments);
 
-        CreateCoreWebView2Environment(environmentcreatedhandler)
-          .map_err(webview2_com::Error::WindowsError)
+        // if data_directory is None, we set it to a null PWSTR
+        let data_directory: PWSTR = data_directory
+          .map(|s| pwstr_from_str(&s))
+          .unwrap_or_default();
+
+        let result = CreateCoreWebView2EnvironmentWithOptions(
+          PWSTR::default(),
+          data_directory,
+          options,
+          environmentcreatedhandler,
+        )
+        .map_err(webview2_com::Error::WindowsError);
+        let _ = take_pwstr(data_directory);
+
+        return result;
       }),
       Box::new(move |error_code, environment| {
         error_code?;
@@ -216,8 +218,8 @@ impl InnerWebView {
       settings
         .SetAreDevToolsEnabled(false)
         .map_err(webview2_com::Error::WindowsError)?;
-      if attributes.devtool {
-        settings.SetAreDevToolsEnabled(true);
+      if attributes.devtools {
+        let _ = settings.SetAreDevToolsEnabled(true);
       }
 
       let mut rect = RECT::default();
@@ -255,7 +257,7 @@ window.addEventListener('mousemove', (e) => window.chrome.webview.postMessage('_
             args.TryGetWebMessageAsString(&mut js)?;
             let js = take_pwstr(js);
             if js == "__WEBVIEW_LEFT_MOUSE_DOWN__" || js == "__WEBVIEW_MOUSE_MOVE__" {
-              if !window.is_decorated() && window.is_resizable() {
+              if !window.is_decorated() && window.is_resizable() && !window.is_maximized() {
                 use crate::application::{platform::windows::hit_test, window::CursorIcon};
 
                 let mut point = POINT::default();
@@ -623,11 +625,17 @@ window.addEventListener('mousemove', (e) => window.chrome.webview.postMessage('_
     };
   }
 
-  /// Open the web inspector which is usually called dev tool.
-  pub fn devtool(&self) {
-    let _ = unsafe {
-      self.webview.OpenDevToolsWindow();
-    };
+  #[cfg(any(debug_assertions, feature = "devtools"))]
+  pub fn open_devtools(&self) {
+    let _ = unsafe { self.webview.OpenDevToolsWindow() };
+  }
+
+  #[cfg(any(debug_assertions, feature = "devtools"))]
+  pub fn close_devtools(&self) {}
+
+  #[cfg(any(debug_assertions, feature = "devtools"))]
+  pub fn is_devtools_open(&self) -> bool {
+    false
   }
 }